LPS-71087 Move to new package
