+++ modules/sdk/gradle-plugins-maven-plugin-builder/src/main/java/com/liferay/gradle/plugins/maven/plugin/builder/MavenPluginBuilderPlugin.java
@@ -14,6 +14,7 @@
 
 package com.liferay.gradle.plugins.maven.plugin.builder;
 
+import com.liferay.gradle.plugins.maven.plugin.builder.tasks.BuildPluginDescriptorTask;
 import com.liferay.gradle.util.GradleUtil;
 
 import java.io.File;
