+++ modules/sdk/gradle-plugins-maven-plugin-builder/src/main/java/com/liferay/gradle/plugins/maven/plugin/builder/tasks/BuildPluginDescriptorTask.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.gradle.plugins.maven.plugin.builder;
+package com.liferay.gradle.plugins.maven.plugin.builder.tasks;
 
 import com.liferay.gradle.plugins.maven.plugin.builder.internal.util.XMLUtil;
 import com.liferay.gradle.util.GradleUtil;
